:recycle: Rename DelegatesV2 component to Delegates
