+++ src/components/delegates/index.js
@@ -1,7 +1,7 @@
 /* istanbul ignore file */
 import { connect } from 'react-redux';
 import { translate } from 'react-i18next';
-import DelegatesV2 from './delegatesV2';
+import Delegates from './delegates';
 import { clearVotes } from '../../actions/voting';
 
 const mapStateToProps = state => ({
@@ -12,4 +12,4 @@ const mapDispatchToProps = {
   clearVotes,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(translate()(DelegatesV2));
+export default connect(mapStateToProps, mapDispatchToProps)(translate()(Delegates));
