+++ src/components/delegates/votingHeader.css
