+++ src/components/delegates/votingHeader.js
